+++ tests/testthat/test_misc_influence.r
@@ -154,3 +154,16 @@ test_that("influence() correctly works with 'na.omit' and 'na.pass'.", {
    expect_equivalent(sum(is.na(sav$dfbs$intrcpt)), 4)
 
 })
+
+test_that("'infonly' argument works correctly with influence().", {
+
+   data(dat.bcg, package="metafor")
+   dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg, slab=paste0("Trial ", dat.bcg$trial))
+   res <- rma(yi, vi, data=dat, method="FE")
+   sav <- influence(res)
+   tmp <- print(sav)
+   expect_equivalent(length(tmp$rstudent), 13)
+   tmp <- print(sav, infonly=TRUE)
+   expect_equivalent(length(tmp$rstudent), 3)
+
+})